src/utilities/requirements.txt
packaging==25.0
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.4
